Slovak Translation
I have extensive work experience as a freelance translator SK-EN and I work now as a programmer, so the translation is accurate.

I have just one comment to add - I used direct form of speech, in Slovak it is possible to use a more polite version of speaking, e.g. Usted in spanish or Sie in german, but that took up much more space and I find the direct form of speech much more friendly.

It was a pleasure contributing to this great project :)

P.S. There might be slight issues in some cases because in Slovak plurals can have different forms, e.g. 2-4 Pomodoros are "Pomodorá" and 5+ are "Pomodor", I used appropriate form where possible, that's the best I could do.